requirements.txt
Django
djangorestframework
python-dotenv
openai
PyPDF2
python-docx